cleanup, changed header file ext to .hpp, renamed Main to Test

+++ ServiceLocator.hpp
@@ -2,7 +2,6 @@
 
 #include <functional>
 #include <memory>
-#include <type_traits>
 #include <unordered_map>
 
 class ServiceLocator
@@ -23,8 +22,8 @@ public:
 	std::shared_ptr<T> resolve() const;
 
 private:
-	std::unordered_map<int, std::shared_ptr<void>> instances;
-	std::unordered_map<int, std::function<std::shared_ptr<void>()>> creators;
+	std::unordered_map<size_t, std::shared_ptr<void>> instances;
+	std::unordered_map<size_t, std::function<std::shared_ptr<void>()>> creators;
 };
 
 #include "ServiceLocator.inl"
